split routes from business functions
